chore: move generate out of /src/deprecated
